Core refactor
