+++ Files/Helpers/NativeFileOperationsHelper.cs
@@ -1,0 +1,154 @@
+﻿using System;
+using System.Runtime.InteropServices;
+using System.Runtime.InteropServices.ComTypes;
+
+namespace Files.Helpers
+{
+    public class NativeFileOperationsHelper
+    {
+        public enum File_Attributes : uint
+        {
+            Readonly = 0x00000001,
+            Hidden = 0x00000002,
+            System = 0x00000004,
+            Directory = 0x00000010,
+            Archive = 0x00000020,
+            Device = 0x00000040,
+            Normal = 0x00000080,
+            Temporary = 0x00000100,
+            SparseFile = 0x00000200,
+            ReparsePoint = 0x00000400,
+            Compressed = 0x00000800,
+            Offline = 0x00001000,
+            NotContentIndexed = 0x00002000,
+            Encrypted = 0x00004000,
+            Write_Through = 0x80000000,
+            Overlapped = 0x40000000,
+            NoBuffering = 0x20000000,
+            RandomAccess = 0x10000000,
+            SequentialScan = 0x08000000,
+            DeleteOnClose = 0x04000000,
+            BackupSemantics = 0x02000000,
+            PosixSemantics = 0x01000000,
+            OpenReparsePoint = 0x00200000,
+            OpenNoRecall = 0x00100000,
+            FirstPipeInstance = 0x00080000
+        }
+
+        public const uint GENERIC_READ = 0x80000000;
+
+        [DllImport("api-ms-win-core-file-fromapp-l1-1-0.dll", CharSet = CharSet.Auto,
+        CallingConvention = CallingConvention.StdCall,
+        SetLastError = true)]
+        public static extern IntPtr CreateFileFromApp(
+            string lpFileName,
+            uint dwDesiredAccess,
+            uint dwShareMode,
+            IntPtr SecurityAttributes,
+            uint dwCreationDisposition,
+            uint dwFlagsAndAttributes,
+            IntPtr hTemplateFile
+        );
+
+        [DllImport("api-ms-win-core-file-fromapp-l1-1-0.dll", CharSet = CharSet.Auto,
+        CallingConvention = CallingConvention.StdCall,
+        SetLastError = true)]
+        public static extern IntPtr CreateFile2FromApp(
+            string lpFileName,
+            uint dwDesiredAccess,
+            uint dwShareMode,
+            uint dwCreationDisposition,
+            IntPtr pCreateExParams
+        );
+
+        [DllImport("api-ms-win-core-file-fromapp-l1-1-0.dll", CharSet = CharSet.Auto,
+        CallingConvention = CallingConvention.StdCall,
+        SetLastError = true)]
+        public static extern bool MoveFileFromApp(
+            string lpExistingFileName,
+            string lpNewFileName
+        );
+
+        [DllImport("api-ms-win-core-file-fromapp-l1-1-0.dll", CharSet = CharSet.Auto,
+        CallingConvention = CallingConvention.StdCall,
+        SetLastError = true)]
+        public static extern bool CopyFileFromApp(
+            string lpExistingFileName,
+            string lpNewFileName,
+            bool bFailIfExists
+        );
+
+        [DllImport("api-ms-win-core-file-fromapp-l1-1-0.dll", CharSet = CharSet.Auto,
+        CallingConvention = CallingConvention.StdCall,
+        SetLastError = true)]
+        public static extern bool DeleteFileFromApp(
+            string lpFileName
+        );
+
+        [DllImport("api-ms-win-core-file-fromapp-l1-1-0.dll", CharSet = CharSet.Auto,
+        CallingConvention = CallingConvention.StdCall,
+        SetLastError = true)]
+        public static extern bool RemoveDirectoryFromApp(
+            string lpPathName
+        );
+
+        [DllImport("api-ms-win-core-file-fromapp-l1-1-0.dll", SetLastError = true, CharSet = CharSet.Auto)]
+        [return: MarshalAs(UnmanagedType.Bool)]
+        public static extern bool GetFileAttributesExFromApp(
+            string lpFileName,
+            GET_FILEEX_INFO_LEVELS fInfoLevelId,
+            out WIN32_FILE_ATTRIBUTE_DATA lpFileInformation);
+
+        [DllImport("api-ms-win-core-file-fromapp-l1-1-0.dll", SetLastError = true, CharSet = CharSet.Auto)]
+        [return: MarshalAs(UnmanagedType.Bool)]
+        public static extern bool SetFileAttributesFromApp(
+            string lpFileName,
+            System.IO.FileAttributes dwFileAttributes);
+
+        public enum GET_FILEEX_INFO_LEVELS
+        {
+            GetFileExInfoStandard,
+        }
+
+        [StructLayout(LayoutKind.Sequential)]
+        public struct WIN32_FILE_ATTRIBUTE_DATA
+        {
+            public System.IO.FileAttributes dwFileAttributes;
+            public FILETIME ftCreationTime;
+            public FILETIME ftLastAccessTime;
+            public FILETIME ftLastWriteTime;
+            public uint nFileSizeHigh;
+            public uint nFileSizeLow;
+        }
+
+        public static bool HasFileAttribute(string lpFileName, System.IO.FileAttributes dwAttrs)
+        {
+            if (GetFileAttributesExFromApp(
+                lpFileName, GET_FILEEX_INFO_LEVELS.GetFileExInfoStandard, out var lpFileInfo))
+            {
+                return (lpFileInfo.dwFileAttributes & dwAttrs) == dwAttrs;
+            }
+            return false;
+        }
+
+        public static bool SetFileAttribute(string lpFileName, System.IO.FileAttributes dwAttrs)
+        {
+            if (!GetFileAttributesExFromApp(
+                lpFileName, GET_FILEEX_INFO_LEVELS.GetFileExInfoStandard, out var lpFileInfo))
+            {
+                return false;
+            }
+            return SetFileAttributesFromApp(lpFileName, lpFileInfo.dwFileAttributes | dwAttrs);
+        }
+
+        public static bool UnsetFileAttribute(string lpFileName, System.IO.FileAttributes dwAttrs)
+        {
+            if (!GetFileAttributesExFromApp(
+                lpFileName, GET_FILEEX_INFO_LEVELS.GetFileExInfoStandard, out var lpFileInfo))
+            {
+                return false;
+            }
+            return SetFileAttributesFromApp(lpFileName, lpFileInfo.dwFileAttributes & ~dwAttrs);
+        }
+    }
+}
